Update validate & delete, Add temp_function

+++ public/application/class/Ad.php
@@ -93,9 +93,18 @@ class Ad {
      */
     static function delete($slug){
 
-        $query= "DELETE FROM ca_ad WHERE a_unique_id = :uniqueId";
+        $query= "SELECT a_id FROM ca_ad WHERE a_unique_id = :uniqueId";
 
-        \classified_ads\Bdd::executeSql($query,[':uniqueId'=>hash('sha1',$slug)],[':uniqueId'=>PDO::PARAM_STR]);
+        $response = \classified_ads\Bdd::executeSql($query,[':uniqueId'=>hash('sha1',$slug)],[':uniqueId'=>PDO::PARAM_STR]);
+        if($response->fetch(PDO::FETCH_ASSOC)){
+            $query= "DELETE FROM ca_ad WHERE a_unique_id = :uniqueId";
+
+            \classified_ads\Bdd::executeSql($query,[':uniqueId'=>hash('sha1',$slug)],[':uniqueId'=>PDO::PARAM_STR]);
+            return true;
+        }else{
+            return false;
+        }
+        
 
     }
 
@@ -104,31 +113,39 @@ class Ad {
      * avec envoi de mail
      *
      * @param [String] $slug - annonce à valider
-     * @return void
+     * @return [Boolean]
      */
     static function validate($slug) {
 
         $mail = \classified_ads\Crypt::decryptSimple(explode('&',$slug)[0]);
         $id = \classified_ads\Crypt::decryptSimple(explode('&',$slug)[1]);
 
-        $mail_crypt = \classified_ads\Crypt::encryptSimple($mail);
-        $id_crypt = \classified_ads\Crypt::encryptSimple($id);
+        $query = "SELECT a_id FROM ca_ad INNER JOIN ca_user ON ca_ad.a_u_id = ca_user.u_id WHERE a_unique_id = :id AND u_mail = :mail ";
+        $response = \classified_ads\Bdd::executeSql($query,[':id'=>hash('sha1',$slug),':mail'=>$mail],[':id'=>PDO::PARAM_INT,':mail'=>PDO::PARAM_STR]);
 
-        $new = hash('sha1',"$mail_crypt&$id_crypt");
+        if($response->fetch(PDO::FETCH_ASSOC)){
+            
+            $mail_crypt = \classified_ads\Crypt::encryptSimple($mail);
+            $id_crypt = \classified_ads\Crypt::encryptSimple($id);
+    
+            $new = hash('sha1',"$mail_crypt&$id_crypt");
+    
+            $query = "UPDATE ca_ad INNER JOIN ca_user ON ca_ad.a_u_id = ca_user.u_id SET a_validate = :validate, a_date_validate = :date, a_unique_id = :newUnique WHERE a_unique_id = :uniqueId AND u_mail = :mail AND a_id = :id ";
+    
+            $param = [':validate'=>true,':date'=>date('Y-m-d'),':newUnique'=>$new,':uniqueId'=>hash('sha1',$slug),':mail'=>$mail,':id'=>$id];
+    
+            $type = [':validate'=>PDO::PARAM_BOOL,':date'=>PDO::PARAM_STR,':newUnique'=>PDO::PARAM_STR,':uniqueId'=>PDO::PARAM_STR,':mail'=>PDO::PARAM_STR,':id'=>PDO::PARAM_INT];
+            
+            \classified_ads\Bdd::executeSql($query,$param,$type)->fetch();
+                // \classified_ads\Ad::updateId(hash('sha1',$slug),$new);
+            $message = "$mail_crypt\n$id_crypt ----------- <a href = '".SERVER_URI."/delete-$mail_crypt&$id_crypt>suppression </a>\n edition: <a href = '".SERVER_URI."/edit-$mail_crypt&$id_crypt'>edition</a>";
+    
+            mail($mail,"Validation de votre annonce",$message);
 
-        $query = "UPDATE ca_ad INNER JOIN ca_user ON ca_ad.a_u_id = ca_user.u_id SET a_validate = :validate, a_date_validate = :date, a_unique_id = :newUnique WHERE a_unique_id = :uniqueId AND u_mail = :mail AND a_id = :id ";
-
-        $param = [':validate'=>true,':date'=>date('Y-m-d'),':newUnique'=>$new,':uniqueId'=>hash('sha1',$slug),':mail'=>$mail,':id'=>$id];
-
-        $type = [':validate'=>PDO::PARAM_BOOL,':date'=>PDO::PARAM_STR,':newUnique'=>PDO::PARAM_STR,':uniqueId'=>PDO::PARAM_STR,':mail'=>PDO::PARAM_STR,':id'=>PDO::PARAM_INT];
-
-        \classified_ads\Bdd::executeSql($query,$param,$type);
-
-        \classified_ads\Ad::updateId(hash('sha1',$slug),$new);
-        $message = "$mail_crypt\n$id_crypt ----------- <a href = '".SERVER_URI."/delete-$mail_crypt&$id_crypt>suppression </a>\n edition: <a href = '".SERVER_URI."/edit-$mail_crypt&$id_crypt'>edition</a>";
-
-        mail($mail,"Validation de votre annonce",$message);
-
+            return true;
+        }else{
+            return false;
+        }
     }
 
     /**
@@ -154,5 +171,11 @@ class Ad {
         $param = [':uniqueId'=>$new,':id'=>$id];
         $type = [':uniqueId'=>PDO::PARAM_STR,':id'=>PDO::PARAM_STR];
         \classified_ads\Bdd::executeSql($query,$param,$type);
+    }
+
+    static function abricot($slug){
+
+        $query = "SELECT * FROM ca_ad INNER JOIN ca_user ON a_u_id = u_id WHERE a_unique_id = :id";
+        return \classified_ads\Bdd::executeSql($query,[':id'=>hash('sha1',$slug)],[':id'=>PDO::PARAM_STR])->fetchALL(PDO::FETCH_ASSOC);
     }
 }
